Adding team progress widget
Due to the similarity of the code here, I made this the same html page
as the individual progress one. The url var 'team' can be set to true
to use it for a team, or false (or left out) to use it for an individual.
I renamed the widget to reflect the new behavior.